[file-index] Support V2 file index container format

+++ paimon-common/src/main/java/org/apache/paimon/fileindex/FileIndexer.java
@@ -32,7 +32,7 @@ public interface FileIndexer {
 
     FileIndexWriter createWriter();
 
-    FileIndexReader createReader(SeekableInputStream inputStream, int start, int length);
+    FileIndexReader createReader(SeekableInputStream inputStream, long start, int length);
 
     static FileIndexer create(String type, DataType dataType, Options options) {
         FileIndexerFactory fileIndexerFactory = FileIndexerFactoryUtils.load(type);
