+++ paimon-common/src/main/java/org/apache/paimon/fileindex/bitmap/BitmapFileIndex.java
@@ -69,7 +69,7 @@ public class BitmapFileIndex implements FileIndexer {
 
     @Override
     public FileIndexReader createReader(
-            SeekableInputStream seekableInputStream, int start, int length) {
+            SeekableInputStream seekableInputStream, long start, int length) {
         try {
             return new Reader(seekableInputStream, start, options);
         } catch (Exception e) {
@@ -193,7 +193,7 @@ public class BitmapFileIndex implements FileIndexer {
     private static class Reader extends FileIndexReader {
 
         private final SeekableInputStream seekableInputStream;
-        private final int headStart;
+        private final long headStart;
         private final Map<Object, RoaringBitmap32> bitmaps = new LinkedHashMap<>();
 
         private BitmapFileIndexMeta bitmapFileIndexMeta;
@@ -201,7 +201,7 @@ public class BitmapFileIndex implements FileIndexer {
 
         private final Options options;
 
-        public Reader(SeekableInputStream seekableInputStream, int start, Options options) {
+        public Reader(SeekableInputStream seekableInputStream, long start, Options options) {
             this.seekableInputStream = seekableInputStream;
             this.headStart = start;
             this.options = options;
