+++ paimon-common/src/main/java/org/apache/paimon/fileindex/bloomfilter/BloomFilterFileIndex.java
@@ -69,7 +69,7 @@ public class BloomFilterFileIndex implements FileIndexer {
     }
 
     @Override
-    public FileIndexReader createReader(SeekableInputStream inputStream, int start, int length) {
+    public FileIndexReader createReader(SeekableInputStream inputStream, long start, int length) {
         try {
             inputStream.seek(start);
             byte[] serializedBytes = new byte[length];
